cambio de estructura
