+++ scss/global.scss
@@ -5,6 +5,7 @@
 @import "../components/navbar/navbar.scss";
 @import "../pages/login/login.scss";
 @import "../pages/dashboard/dashboard.scss";
+@import "../pages/modCarga/modCarga.scss";
 @import "../pages/modAcciones/modAcciones.scss";
 
 body{
